Remove lookup services from common and move to datastore, build out lookup service and interface initial structure
