+++ modules/common/drush.services.yml
@@ -5,7 +5,6 @@ services:
       - { name: drush.command }
     arguments:
       - '@dkan.common.dataset_info'
-      - '@database'
 
   common.jobstore.commands:
     class: \Drupal\common\Commands\JobStoreCommands
